[Event Creation] - Memory optimization
I was storing the whole list of events in an array. That won't be okay
when we have a lot of events. Now I just query for the name of each
event while I compare the names. Also, I only update the list of hosted
events when the user tries to add the event.