[Event Hosting] - add button to end event
This button changes the active field in the event to 0 and it only
appears if the current user is the host. To facilitate, I added
stateParameters to a few states that pass around the user's email. I
couldn't find a way to pass the parameters to the Settings state, but I
just let it lose the user email and I refetch it when I need it