Add the ability to like pictures
When images are uploaded, the number of likes is set to 0 and the list
of users that have liked the photo contains only the uploader of the
picture set to 0 likes (firebase does not allow for empty lists which
would make more sense here, so this circumvents the problem).
As of now, the user can only like the photo but cannot unlike it. I was
having performance issues, so feature is not included in this commit.
Modified eventPageController.js and eventPage.html.